feat: Get CPU Usage through sar Unix command
Calculate CPU Usage through /proc/stat file showed very inappropriated. First, the concept of CPU Usage go further
than take this information from time to time but it require a better understanding from processes states and how OS handles with it.
In order to be fast and precise will be used the 'sar' command line which is able to show the %idle time for every core.

The challenge here is to get this information from a file which is used to pass the result from 'sar' command.

Issue: #1