REFACTOR %CPU
The function to see CPU Usage was modified. Before it was used the /proc/stats geting the time spent by OS in the respective
process state. This measurement didn't provide good information to work with %CPU. Therefore, the strategy now is to use
'sar' Unix Command, which is a system's activity collector and reporter. It requires some packages to work with it, but one command
to install all them.
Here it's getting the output from 'sar -P ALL 1 1' command, which perform a report to every core, in 1 second, the fraction of time
that it's spending in every process state. This information is sent to a file and the code get the %idle in every core. This process
still isn't so general.
It generate a bug in terminal colors.

Resolves: #1